Separated the model and the service for accessing Wikidata and Mongo.
